Small Refactor
